#82 #141 Moved code for Quiz History Database to OpenActionsFragment
